requirements.txt
Flask==2.0.3
flask-restx==0.5.1
pymongo==4.1.1
python-dotenv==0.20.0
bcrypt==3.2.0
marshmallow==3.14.1
pytest==7.1.2
Werkzeug==2.0.3